Crud veterinario terminado

+++ veterinariaapp/views.py
@@ -1,30 +1,42 @@
 from django.http import HttpResponse
-from django.shortcuts import render
-from .forms import VeterinarioFrom
+from django.shortcuts import render, redirect
+from .forms import VeterinarioForm
 from .models import *
 from django.contrib.auth.models import User
 
-#crear vista home
+#VISTA DEL HOME PAGE
 def home(request):
     return render(request, 'home/index.html')
 
-#tabla de los datos del o de los 
+#CRUD DEL MODELO DEL VETERINARIO
+#TABLA DEL MODELO DE VETERINARIO
 def lista_veterinario(request):
     veterinarios= Veterinario.objects.all().order_by('id')
     return render(request, 'veterinario/lista_veterinario.html',{'veterinarios':veterinarios})
 
-# Crear las vistas aqui 
-
+#LOGICA BASA EN FUNCIONES DE CREAR DEL MODELO DE VETERINARIO
 def crear_veterinario(request):
     if request.method == 'POST':
-        form = VeterinarioFrom(request.POST)
+        form = VeterinarioForm(request.POST)
         if form.is_valid():
-            usuario = form.cleaned_data['usuario']
+            username = form.cleaned_data['username']
             password = form.cleaned_data['password']
+            email=form.cleaned_data['email']
+            nombres=form.cleaned_data['nombres']
+            apellidos=form.cleaned_data['apellidos']
+            tipo_documento=form.cleaned_data['tipo_documento']
+            num_documento= form.cleaned_data['num_documento']
+            celular= form.cleaned_data['celular']
             veterinario=form.save(commit=False)
             persona = Persona.objects.create_user(
-                username=usuario, 
-                password=password
+                username=username, 
+                password=password,
+                email=email,
+                first_name=nombres,
+                last_name=apellidos,
+                tipo_documento=tipo_documento,
+                num_documento=num_documento,
+                celular=celular
                 )
             veterinario.persona = persona
             veterinario.save()
@@ -35,16 +47,14 @@ def crear_veterinario(request):
             mensaje=f'El veterinario {veterinario} fue agreado'
             return render(request, 'veterinario/mensaje.html',{'mensaje':mensaje})
     else:
-        form= VeterinarioFrom()
+        form= VeterinarioForm()
         return render (request,'veterinario/crear_veterinario.html',{'form':form})
-    
 
-
-#Actualizar CRUD veterinario
+#LOGICA BASA EN FUNCIONES DE ACTUALIZAR DEL MODELO DE VETERINARIO
 def actualizar_veterinario(request, id):
     veterinario= Veterinario.objects.get(id=id)
     if request.method == 'POST':
-        form = VeterinarioFrom(request.POST, instance=veterinario)
+        form = VeterinarioForm(request.POST, instance=veterinario)
         if form.is_valid():
             veterinario= form.save()
             mensaje= f'El veterinario {veterinario} fue actualizado correctamente'
@@ -52,12 +62,14 @@ def actualizar_veterinario(request, id):
             mensaje= f'El veterinario {veterinario} no fue actualizado'
         return render (request,'veterinario/mensaje.html',{'mensaje':mensaje})
     else:
-        form = VeterinarioFrom(instance=veterinario)
+        form = VeterinarioForm(instance=veterinario)
         return render(request, 'veterinario/crear_veterinario.html',{'form':form})
 
-#eliminar -CRUD- veterinario
+#LOGICA BASA EN FUNCIONES DE ELIMINAR DEL MODELO DE VETERINARIO
 def eliminar_veterinario(request, id):
     veterinario= Veterinario.objects.get(id=id)
-    veterinario.delete()
-    mensaje= f'El veterinario {veterinario} fue eliminado correctamente'
+    if request.method == 'POST':
+        veterinario.delete()
+        mensaje= f'El veterinario {veterinario} fue eliminado correctamente'
+        return redirect('veterinario/lista_veterinario.html')
     return render(request, 'veterinario/eliminar_veterinario.html',{'veterinario':veterinario})
